server/requirements.txt
flask==3.1.0
flask-cors==4.0.0
python-dotenv==1.0.1
requests==2.31.0
azure-cognitiveservices-speech==1.46.0
openai==1.14.0
azure-identity==1.15.0
gunicorn==21.2.0